major revision

+++ backend/app.py
@@ -8,26 +8,38 @@ from sqlite3 import connect, Row
 from uuid import uuid4
 from werkzeug.exceptions import HTTPException
 
-def get_entry(uuid: str="") -> dict:
+def get_entry(json: dict) -> dict:
     try:
         with connect("./static/db_final.db") as database_handle:
             database_handle.row_factory = Row
             cursor = database_handle.cursor()
-            if uuid:
+            if json:
+                update_key = tuple(json.keys())[0]
+                if update_key not in (
+                    "uuid", "name", "email", "password", "dob", "address", "phone", "image"):
+                        raise Exception(
+                            "Field with name '%s' not found."%(update_key))
                 cursor.execute("""
-                    SELECT * FROM randomdb WHERE uuid=\"%s\"
-                """%(uuid))
+                    SELECT * FROM randomdb WHERE \"%s\"=\"%s\" LIMIT 1
+                """%(update_key, json[update_key]))
             else:
                 cursor.execute("""
                     SELECT * FROM randomdb ORDER BY RANDOM() LIMIT 1
                 """)
             data = cursor.fetchone()
-        return dict(data)
+        if data:
+            return dict(data)
+        else:
+            raise Exception(
+                "Entry with field '%s' and data '%s' not found."%(update_key, json[update_key]))
     except Exception as error:
         print("Error:", error)
-        return {"get": False}
+        return {
+            "get": False,
+            "description": str(error)
+        }
 
-def ins_entry(json: dict) -> bool:
+def ins_entry(json: dict) -> dict:
     try:
         uuid = str(uuid4())
         with connect("./static/db_final.db") as database_handle:
@@ -46,38 +58,63 @@ def ins_entry(json: dict) -> bool:
                 json["image"]
             ))
             database_handle.commit()
-        print("Info:", uuid)
-        return True
+        return {
+            "post": True,
+            "uuid": uuid
+        }
     except Exception as error:
         print("Error:", error)
-        return False
+        return {
+            "post": False,
+            "description": str(error)
+        }
 
-def upd_entry(json: dict) -> bool:
+def upd_entry(json: dict) -> dict:
     try:
         update_key = tuple(json.keys())[1]
+        if update_key not in (
+            "name", "email", "password", "dob", "address", "phone", "image"):
+                raise Exception(
+                    "Field with name '%s' not found."%(update_key))
         with connect("./static/db_final.db") as database_handle:
-            database_handle.row_factory = Row
             cursor = database_handle.cursor()
+            cursor.execute("""
+                    SELECT * FROM randomdb WHERE uuid=\"%s\"
+            """%(json["uuid"]))
+            if not cursor.fetchone():
+                raise Exception("Entry with UUID '%s' not found."%(json["uuid"]))
             cursor.execute("""
                 UPDATE randomdb SET %s=\"%s\" WHERE uuid=\"%s\"
             """%(update_key, json[update_key], json["uuid"]))
             database_handle.commit()
-        return True
+        return {"patch": True}
     except Exception as error:
         print("Error:", error)
-        return False
+        return {
+            "patch": False,
+            "error": str(error)
+        }
 
-def del_entry(identifier: str) -> bool:
+def del_entry(uuid: str) -> dict:
     try:
         with connect("./static/db_final.db") as database_handle:
-            database_handle.cursor().execute("""
+            cursor = database_handle.cursor()
+            cursor.execute("""
+                    SELECT * FROM randomdb WHERE uuid=\"%s\"
+            """%(uuid))
+            if not cursor.fetchone():
+                raise Exception("Entry with UUID '%s' not found."%(uuid))
+            cursor.execute("""
                 DELETE FROM randomdb WHERE uuid=\"%s\"
-            """%(identifier))
+            """%(uuid))
             database_handle.commit()
-        return True
+        return {"delete": True}
     except Exception as error:
         print("Error:", error)
-        return False
+        return {
+            "delete": False,
+            "description": str(error)
+        }
 
 app = Flask(__name__)
 
@@ -87,26 +124,23 @@ def root():
 
 @app.route("/get", methods=["GET"])
 def get_handle() -> dict:
-    return get_entry()
+    return get_entry({})
 
-@app.route("/query/", methods=["GET"])
-def query_handle() -> dict:
-    if request.args.get("uuid"):
-        return get_entry(request.args.get("uuid"))
-    else:
-        return  {"query": None}
+@app.route("/search", methods=["POST"])
+def search_handle() -> dict:
+    return get_entry(request.get_json())
 
 @app.route("/post", methods=["POST"])
 def post_handle() -> dict:
-    return {"post": ins_entry(request.get_json())}
+    return ins_entry(request.get_json())
 
 @app.route("/patch", methods=["PATCH"])
 def patch_handle() -> dict:
-    return {"patch": upd_entry(request.get_json())}
+    return upd_entry(request.get_json())
 
-@app.route("/delete", methods=["DELETE"])
+@app.route("/delete", methods=["POST"])
 def delete_handle() -> dict:
-    return {"delete": del_entry(request.get_json()["uuid"])}
+    return del_entry(request.get_json()["uuid"])
 
 @app.errorhandler(HTTPException)
 def error_handle(error):
